+++ java/effective-java/note-app/memo/src/main/java/me/staek/memo/EditMenu.java
@@ -1,6 +1,6 @@
 package me.staek.memo;
 
-import me.staek.memo.menuenum.Menu;
+import me.staek.memo.code.Menu;
 
 public class EditMenu implements MemoMenu {
     MemoFrame memoFrame;
