+++ java/effective-java/note-app/memo/src/main/java/me/staek/memo/FileMenu.java
@@ -1,7 +1,5 @@
 package me.staek.memo;
 
-import me.staek.memo.menuenum.Menu;
-
 import java.awt.*;
 import java.io.BufferedReader;
 import java.io.FileReader;
