+++ java/effective-java/note-app/memo/src/main/java/me/staek/memo/code/MenuType.java
@@ -1,4 +1,4 @@
-package me.staek.memo.menuenum;
+package me.staek.memo.code;
 
 /**
  * menu type
