+++ java/effective-java/note-app/memo/src/main/java/me/staek/memo/listener/FontMenuListener.java
@@ -3,7 +3,7 @@ package me.staek.memo.listener;
 
 import me.staek.memo.handler.MemoActionHandler;
 import me.staek.memo.MemoMenu;
-import me.staek.memo.menuenum.Menu;
+import me.staek.memo.code.Menu;
 
 import java.awt.event.ActionListener;
