+++ java/effective-java/note-app/memo/src/test/java/me/staek/memo/code/MenuItemEnumTest.java
@@ -1,4 +1,4 @@
-package me.staek.memo.menuenum;
+package me.staek.memo.code;
 
 import org.junit.jupiter.api.Test;
